fix: create `Back` button only in case of less then 4 columns
This pull request fixes an issue where the Back and Navigation buttons were only created, when there were more than 4 columns or 12 buttons. The Back button is now created when there are less than 4 columns or 12 buttons. This ensures that the Back button is displayed in case of fewer buttons.

Additional Information:
The fix is a simple change that has been tested.

I hope this helps!